Move redeployment when bootsrapping logic to boostrap class
